+++ userscript/script.safari.js
@@ -1,0 +1,2 @@
+import('https://userscript.scratchaddons.cf/userscript/script.main.js')
+completion()
